Add Peer-To-Peer priority mechanism
Most of NFC devices support card emulation now, there is a chance
that the remote NFC device is detected as a card not peer device.
This peer-to-peer priority mechanism will always try to establish
peer-to-peer connection than card emulation

Peer-To-Peer priority mechanism works only if NFCC is running in
bail out mode. The configuration option NFA_POLL_BAIL_OUT_MODE is to
define the mode of NFCC.

Change-Id: I45a099efa77fab132189ea92da97a90240cfba1d